Create DAY 13(FIND LARGEST OF THREE NUMBERS)
We can  compare all the three numbers by using the ternary operator in a single statement. If we want to compare three numbers in a single statement, we must use the following statement.
d = c > (a>b ? a:b) ? c:((a>b) ? a:b);  
In this program, we have used a single statement to find the largest of three numbers.