Update DAY 14 (SUM OF NATURAL NUMBERS)
As we know natural numbers contain all the positive numbers starting from 1, 2, 3, to n numbers or infinity. For example, suppose when we calculate the sum of the first 25 numbers. That means we start adding the numbers from 1 to the given number 25, and the process is called the sum of the first N natural number. Here, we will learn how to find the sum of first n numbers