Create DAY 16(DISPLAY EVEN NUMBERS BETWEEN 1 TO100)
We have declared a variable named number and initialized it with 100 (the limit to print the even number). We have used a for loop
that executes up to 100 times and for each iteration of i the if statement checks the number is even or not. After printing each even number, the value if i is increased by 1.In order to check the number, we have divided the number by 2 if it does not leave any remainder, the number is even and the print statement prints that number.